Signed-off-by: Catayao56 <Catayao56@gmail.com>
Update v0.0.8.0-Linux

Our program is now in the second stage of development!
From developing weekly builds, we moved to the Unstable build!
Sounds not good, but in other words, it is in debugging phase.

We are doing our best to make Shadow Suite better, for novices
and experienced, and take note, FOR WHITE HATS ONLY.

Have a nice day and thank you very much!

-Catayao56